requirements.txt
absl-py==2.1.0
alembic==1.13.1
altair==5.3.0
annotated-types==0.6.0
appdirs==1.4.4
attrs==23.2.0
Authlib==1.3.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
contextlib2==21.6.0
contourpy==1.2.1
cplex==22.1.1.2
cryptography==42.0.5
cycler==0.12.1
dill==0.3.8
dimod==0.12.14
diskcache==5.6.3
docplex==2.27.239
dwave-cloud-client==0.11.3
dwave-greedy==0.3.0
dwave-hybrid==0.6.11
dwave-inspector==0.4.4
dwave-neal==0.6.0
dwave-networkx==0.8.14
dwave-ocean-sdk==6.9.0
dwave-preprocessing==0.6.5
dwave-samplers==1.2.0
dwave-system==1.23.0
dwave-tabu==0.5.0
dwavebinarycsp==0.3.0
fasteners==0.19
Flask==3.0.3
fonttools==4.51.0
frozendict==2.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
homebase==1.0.1
html5lib==1.1
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.2.1
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
minorminer==0.2.13
ml-collections==0.1.1
mpmath==1.3.0
multitasking==0.0.11
networkx==3.3
numpy==1.26.4
optuna==3.6.1
packaging==24.0
pandas==2.2.1
patsy==0.5.6
pbr==6.0.0
peewee==3.17.1
penaltymodel==1.1.0
pillow==10.3.0
plotly==5.20.0
plucky==0.4.3
protobuf==4.25.3
pyarrow==15.0.2
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
qiskit==1.0.2
qiskit-algorithms==0.3.0
qiskit-optimization==0.6.1
rectangle-packer==2.0.2
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rustworkx==0.14.2
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.29
statsmodels==0.14.1
stevedore==5.2.0
streamlit==1.33.0
symengine==0.11.0
sympy==1.12
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
webencodings==0.5.1
Werkzeug==3.0.2
yfinance==0.2.37
zipp==3.18.1
